requirements.txt
pytest==8.1.1
pytest-html==4.1.1
selenium==4.19.0
python-dotenv==1.0.1
requests==2.31.0


